chore: refactor account styling
