+++ packages/pilot/src/containers/Account/PasswordRecovery/Reset/Confirmation.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Confirmation from '../Confirmation'
+import Confirmation from '../../Confirmation'
 
 const ResetConfirmation = ({
   onBackToLogin,
