+++ packages/pilot/src/containers/Account/SignUp/Confirmation.js
@@ -1,45 +1,24 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import classNames from 'classnames'
-import { Button } from 'former-kit'
+import Confirmation from '../Confirmation'
 
-import styles from '../style.css'
-
-const Confirmation = ({
+const SignupConfirmation = ({
   onBackToLogin,
   t,
 }) => (
-  <div
-    className={classNames(
-      styles.primaryContent,
-      styles.confirmationContent
-    )}
-  >
-    <div className={styles.login}>
-      <p className={styles.paragraph}>
-        <b>{t('sign_up.confirmation_emphasis')}</b>
-        {' '}
-        <span>{t('sign_up.confirmation')}</span>
-      </p>
-    </div>
-    <div className={styles.actions}>
-      <div className={styles.hugeButton}>
-        <Button
-          type="button"
-          size="huge"
-          fill="gradient"
-          onClick={onBackToLogin}
-        >
-          {t('back_login_action')}
-        </Button>
-      </div>
-    </div>
-  </div>
+  <Confirmation
+    labels={{
+      backToLogin: t('back_login_action'),
+      confirmation: t('sign_up.confirmation'),
+      confirmationEmphasis: t('sign_up.confirmation_emphasis'),
+    }}
+    onBackToLogin={onBackToLogin}
+  />
 )
 
-Confirmation.propTypes = {
+SignupConfirmation.propTypes = {
   onBackToLogin: PropTypes.func.isRequired,
   t: PropTypes.func.isRequired,
 }
 
-export default Confirmation
+export default SignupConfirmation
